Adding types file to rescanner folder
